requirements.txt
attrs==21.4.0
Babel==2.9.1
cffi==1.15.0
chardet==4.0.0
click==8.1.2
cryptography==36.0.2
importlib-metadata==4.11.3
importlib-resources==5.7.0
iso4217==1.9.20220401
money==1.3.0
pdfminer.six==20220319
pycparser==2.21
python-decouple==3.6
pytz==2022.1
typing_extensions==4.1.1
zipp==3.8.0
